+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @mirrorworld/mirage.core
 
+## 1.0.5
+
+### Patch Changes
+
+- Fixes exports field
+
 ## 1.0.4
 
 ### Patch Changes