+++ packages/frameworks/vue/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @mirrorworld/mirage.vue
 
+## 0.0.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @mirrorworld/mirage.core@1.0.5
+
 ## 0.0.6
 
 ### Patch Changes